THRIFT-385. hs: 64-bit integer and double types incorrectly serialized on 32-bit platforms
Use 64-bit types where appropriate.



git-svn-id: https://svn.apache.org/repos/asf/incubator/thrift/trunk@757619 13f79535-47bb-0310-9956-ffa450edef68

+++ compiler/cpp/src/generate/t_hs_generator.cc
@@ -1453,7 +1453,7 @@ string t_hs_generator::render_hs_type(t_type* type) {
     case t_base_type::TYPE_I32:
       return "Int";
     case t_base_type::TYPE_I64:
-      return "Int";
+      return "Int64";
     case t_base_type::TYPE_DOUBLE:
       return "Double";
     }
